feat: add WiFi streaming with JPEG encoding for pedestrian detection
  - Implement HTTP server with MJPEG streaming at ~10 FPS
  - Add WiFi station mode with auto-reconnect
  - Integrate esp32-camera library for RGB565 to JPEG encoding
  - Lower detection threshold to 0.5 for improved multi-person detection
  - Fix RGB565 color palette for proper bounding box rendering
  - Add dependency lock files for reproducible builds

+++ components/who_peripherals/who_cam/who_cam_s3/who_cam_define.hpp
@@ -11,8 +11,8 @@ inline dl::image::img_t fb2img(const who::cam::cam_fb_t *fb)
 {
     assert(fb->format == PIXFORMAT_RGB565 || fb->format == PIXFORMAT_RGB888);
     return {.data = fb->buf,
-            .width = (int)fb->width,
-            .height = (int)fb->height,
+            .width = static_cast<uint16_t>(fb->width),
+            .height = static_cast<uint16_t>(fb->height),
             .pix_type = (fb->format == PIXFORMAT_RGB565) ? dl::image::DL_IMAGE_PIX_TYPE_RGB565
                                                          : dl::image::DL_IMAGE_PIX_TYPE_RGB888};
 }
